requirements.txt
python-dotenv
openai
pyupbit # 업비트 모듈
python-bithumb # 빗썸 모듈